A Sequence panel, lyric import, and the piece the toolkit is for
The editor gains a twelfth panel: one horizontal time axis with a lane for
shots, a lane per actor and a lane for overlays, over a ruler carrying the song's
own sections and beats. Click to scrub, drag a block to move it, drag its edge to
resize, wheel to zoom about the pointer -- and the bake runs on mouse-release,
because a bake rebuilds tracks and layers and doing that sixty times a second
during a drag would make a smooth drag feel like a stutter.

Lyrics import from LRC, SRT and WebVTT, decided by what is in the file rather
than by its extension: people rename these constantly, and a .txt full of
[00:12.40] is an LRC whatever it is called. SRT and WebVTT carry end times and
keep them; LRC has only starts, so a line runs until the next one begins less a
gap, and the last holds.

And the proof-of-concept, which is the part of this that is not a class diagram.
"Night Shift": five shots cut to a generated song's own sections, a recurring
character walking a stylised miniature city from dusk to night, sixteen lyric
lines, a title, a border, and four small audio-reactive relationships. Three
generators build it -- the score, the city and the project -- so the whole piece
can be rebuilt from source, and the result is an ordinary project file that opens
in the editor and needs none of them again.

The city is generated rather than hand-written because a city is a population,
and a population is where procedural instancing earns its keep: one node is a
whole terrace whose buildings differ in height, footprint, tint and how many of
their windows are lit. Two districts share the same ground and are separated only
by visibility -- putting them side by side triples the scene radius, and the
shadow cascades are fitted to that.

+++ src/seq/sequence.cpp
@@ -943,8 +943,8 @@ Result<ShotCamera> cameraFromJson(const json& j) {
         }
     }
     cam.lookAtActor = readString(j, "lookAtActor");
-    cam.lookAtHeight = static_cast<float>(readNumber(j, "lookAtHeight", cam.lookAtHeight));
-    cam.lookAtWeight = static_cast<float>(readNumber(j, "lookAtWeight", cam.lookAtWeight));
+    cam.lookAtHeight = static_cast<float>(readNumber(j, "lookAtHeight", static_cast<double>(cam.lookAtHeight)));
+    cam.lookAtWeight = static_cast<float>(readNumber(j, "lookAtWeight", static_cast<double>(cam.lookAtWeight)));
     return cam;
 }
